+++ extension/src/components/ui/button.tsx
@@ -1,7 +1,7 @@
 import { Slot } from '@radix-ui/react-slot'
 import type { VariantProps } from 'class-variance-authority'
 
-import { cn } from '@/lib/utils'
+import { cn } from '@/libs'
 
 import { buttonVariants } from './button-variants'
