+++ extension/src/components/ui/label.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 
 import * as LabelPrimitive from '@radix-ui/react-label'
 
-import { cn } from '@/lib/utils'
+import { cn } from '@/libs'
 
 function Label({
   className,
