+++ extension/src/components/ui/switch.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 
 import * as SwitchPrimitive from '@radix-ui/react-switch'
 
-import { cn } from '@/lib/utils'
+import { cn } from '@/libs'
 
 function Switch({
   className,
