+++ extension/src/libs/index.ts
@@ -1,1 +1,2 @@
 export { queryClient } from './query-client'
+export { cn } from './utils'
